+++ set_state/lib/screens/clock_page.dart
@@ -1,0 +1,51 @@
+import 'dart:async';
+
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+class ClockPage extends StatelessWidget {
+  const ClockPage({Key key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        appBar: AppBar(
+          title: Text('clock'),
+        ),
+        body: Center(child: Clock()));
+  }
+}
+
+class _ClockState extends State<Clock> {
+  Timer _timer;
+  int time = 0;
+
+  @override
+  void initState() {
+    super.initState();
+    _timer = Timer.periodic(const Duration(seconds: 1), (timer) {
+      setState(() {
+        time = timer.tick;
+      });
+    });
+  }
+
+  @override
+  void dispose() {
+    // TODO: implement dispose
+    super.dispose();
+    _timer.cancel();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Text(time.toString());
+  }
+}
+
+class Clock extends StatefulWidget {
+  Clock({Key key}) : super(key: key);
+
+  @override
+  _ClockState createState() => _ClockState();
+}
